RFCScope/requirements.txt
python-dotenv
openai
tqdm
beautifulsoup4
rapidfuzz
langchain-community
langchain-huggingface
langchain-core
tiktoken
